WIP first pass at more refactors for rattler
